+++ sql/credits.sql
@@ -5,7 +5,7 @@ CREATE TABLE credits (
     address_id       INTEGER UNSIGNED, -- id of record in index_addresses
     -- asset         TEXT,
     asset_id         INTEGER UNSIGNED, -- id of record in assets table
-    quantity         INTEGER UNSIGNED,
+    quantity         BIGINT  UNSIGNED,
     calling_function TEXT,
     -- event         TEXT,
     event_id         INTEGER UNSIGNED  -- id of record in index_transactions
